+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [TD-2495] Added rule implementations endpoint to filter rule implementations
   by rule, rule implementation and dataset structures properties
 - [TD-2624] Support Base64 encoded rule implementation `raw_content` properties
+- [TD-2562] Rule implementations of raw type can have system or structure alias
 
 ## [3.21.1] 2020-05-08